+++ app/components/UI/HardwareWallet/Swaps/hooks/qrScannerUtils.ts
@@ -1,9 +1,5 @@
 import { useCallback, useEffect, useRef } from 'react';
 import { AppState, AppStateStatus } from 'react-native';
-import {
-  type QRHardwareScanError,
-  QRHardwareScanErrorType,
-} from '../../../../../core/HardwareWallet/errors';
 import { QrScanRequestType } from '@metamask/eth-qr-keyring';
 import { SUPPORTED_UR_TYPE } from '../../../../../constants/qr';
 import { MetaMetricsEvents } from '../../../../../core/Analytics';
@@ -33,39 +29,6 @@ interface CameraPermissionRefreshOptions {
 }
 
 /**
- * Builds a flat analytics payload for a QR hardware wallet scan error.
- *
- * @param options - Error details.
- * @param options.error - Human-readable error message.
- * @param options.error_category - Optional categorisation of the error type.
- * @param options.is_ur_format - Whether the scanned QR code used the UR format.
- * @param options.received_ur_type - The UR type string received (included when `error_category` is `WrongURType`).
- * @returns A key-value record suitable for analytics event properties.
- */
-export function buildQrHardwareWalletErrorAnalyticsProperties(options: {
-  error: string;
-  error_category?: QRHardwareScanErrorType;
-  is_ur_format: boolean;
-  received_ur_type?: string;
-}): Record<string, unknown> {
-  const { error, error_category, is_ur_format, received_ur_type } = options;
-  const payload: Record<string, unknown> = {
-    error,
-    is_ur_format,
-  };
-
-  if (error_category !== undefined) {
-    payload.error_category = error_category;
-  }
-
-  if (error_category === QRHardwareScanErrorType.WrongURType) {
-    payload.received_ur_type = received_ur_type ?? '';
-  }
-
-  return payload;
-}
-
-/**
  * Returns the list of expected UR types for the given scan purpose.
  *
  * - `PAIR` → `crypto-hdkey` and `crypto-account`.
@@ -88,7 +51,7 @@ export function getExpectedURTypes(purpose: QrScanRequestType): string[] {
  * Resolves the QR keyring device name on best-effort basis; falls back to
  * `"Unknown"` if the keyring is unavailable.
  *
- * @param properties - Analytics payload (typically from {@link buildQrHardwareWalletErrorAnalyticsProperties}).
+ * @param properties - Analytics payload (typically from {@link buildQrHardwareWalletErrorAnalyticsProperties} in `AnimatedQRScanner.utils`).
  * @param deps - Analytics helpers for tracking events.
  */
 export async function sendQrHardwareErrorAnalytics(
@@ -170,29 +133,4 @@ export function useCameraPermissionRefresh({
       subscription?.remove?.();
     };
   }, [refreshCameraPermission, isActive]);
-}
-
-/**
- * Compares two scan errors for equality based on message and metadata fields.
- *
- * @param previousError - The previous error (or `null` if none).
- * @param currentError - The current error.
- * @returns `true` when both errors carry the same message, type, UR format flag, and received UR type.
- */
-export function isSameScanError(
-  previousError: QRHardwareScanError | null,
-  currentError: QRHardwareScanError,
-): boolean {
-  if (!previousError) {
-    return false;
-  }
-
-  return (
-    previousError.message === currentError.message &&
-    previousError.metadata.qrHardwareScanErrorType ===
-      currentError.metadata.qrHardwareScanErrorType &&
-    previousError.metadata.isUrFormat === currentError.metadata.isUrFormat &&
-    previousError.metadata.receivedUrType ===
-      currentError.metadata.receivedUrType
-  );
 }
