+++ app/components/UI/QRHardware/AnimatedQRScanner.tsx
@@ -49,35 +49,11 @@ import {
   type QRHardwareScanError,
   QRHardwareScanErrorType,
 } from '../../../core/HardwareWallet/errors';
-import { isSameScanError } from './AnimatedQRScanner.utils';
+import {
+  buildQrHardwareWalletErrorAnalyticsProperties,
+  isSameScanError,
+} from './AnimatedQRScanner.utils';
 import usePrevious from '../../../components/hooks/usePrevious';
-
-/**
- * Builds {@link MetaMetricsEvents.HARDWARE_WALLET_ERROR} properties for QR hardware flows.
- *
- * - `error_category` — set for decode / scan pipeline failures (not for native camera errors).
- * - `received_ur_type` — only when `error_category` is `wrong_ur_type` (decoded UR type mismatch).
- * - `is_ur_format` — whether the scanned payload (trimmed) starts with `ur:` (case-insensitive).
- */
-function buildQrHardwareWalletErrorAnalyticsProperties(options: {
-  error: string;
-  error_category?: QRHardwareScanErrorType;
-  is_ur_format: boolean;
-  received_ur_type?: string;
-}): Record<string, unknown> {
-  const { error, error_category, is_ur_format, received_ur_type } = options;
-  const payload: Record<string, unknown> = {
-    error,
-    is_ur_format,
-  };
-  if (error_category !== undefined) {
-    payload.error_category = error_category;
-  }
-  if (error_category === QRHardwareScanErrorType.WrongURType) {
-    payload.received_ur_type = received_ur_type ?? '';
-  }
-  return payload;
-}
 
 const createStyles = (theme: Theme) =>
   StyleSheet.create({
